Add pattern matching to scrape variables
Using python's regex and list of expressions that are used
by bitbake recipe files, and split each line into three parts, if
possible.

The first part should be an uppercase variable name followed by
an underscore, hyphen, bracket, digit, and/or more chars. The
second grouping will be the expression found, and the third is
the value for the variable name. Note that the expression list
must be in order of more complex to least complex, if an '=' is
before a '+=' then the line will be parsed as an equality as
opposed to an string appending.

Removal of the string quotes from the bb file is also incorporated
in the 'clean_values' function.

Unittests added as appropriate.

Signed-off-by: Gabi Beyer <gabib@live.com>